Rearrange authflow test file
