+++ e2e/tests/authflow/cmd.go
@@ -13,29 +13,44 @@ type End2EndCmd struct {
 }
 
 func (e *End2EndCmd) CreateConfigSource() error {
-	cmd := fmt.Sprintf("../dist/authgear-portal internal e2e create-configsource --app-id %s --config-source %s --config-override \"%s\"", e.AppID, e.resolvePath(e.TestCase.AuthgearYAMLSource.Extend), e.TestCase.AuthgearYAMLSource.Override)
+	cmd := fmt.Sprintf(
+		"../dist/authgear-portal internal e2e create-configsource --app-id %s --config-source %s --config-override \"%s\"",
+		e.AppID,
+		e.resolvePath(e.TestCase.AuthgearYAMLSource.Extend),
+		e.TestCase.AuthgearYAMLSource.Override,
+	)
 	return e.execCmd(cmd)
 }
 
 func (e *End2EndCmd) CreateDefaultDomain() error {
-	cmd := fmt.Sprintf("../dist/authgear-portal internal e2e create-default-domain --app-id %s", e.AppID)
+	cmd := fmt.Sprintf(
+		"../dist/authgear-portal internal e2e create-default-domain --app-id %s",
+		e.AppID,
+	)
 	return e.execCmd(cmd)
 }
 
 func (e *End2EndCmd) ImportUsers(jsonPath string) error {
-	cmd := fmt.Sprintf("../dist/authgear internal e2e import-users %s --app-id %s", e.resolvePath(jsonPath), e.AppID)
+	cmd := fmt.Sprintf(
+		"../dist/authgear internal e2e import-users %s --app-id %s",
+		e.resolvePath(jsonPath),
+		e.AppID,
+	)
 	return e.execCmd(cmd)
 }
 
 func (e *End2EndCmd) ExecuteCustomSQL(sqlPath string) error {
-	cmd := fmt.Sprintf("../dist/authgear internal e2e exec-sql --app-id %s --custom-sql \"%s\"", e.AppID, e.resolvePath(sqlPath))
+	cmd := fmt.Sprintf(
+		"../dist/authgear internal e2e exec-sql --app-id %s --custom-sql \"%s\"",
+		e.AppID,
+		e.resolvePath(sqlPath),
+	)
 	return e.execCmd(cmd)
 }
 
 func (e *End2EndCmd) resolvePath(p string) string {
 	return path.Join("./tests/authflow/", path.Dir(e.TestCase.Path), p)
 }
-
 func (e *End2EndCmd) execCmd(cmd string) error {
 	execCmd := exec.Command("sh", "-c", cmd)
 	fmt.Println(execCmd.String())
